Fix to linker scripts.
Placement of the data section in ROM was not always exactly at _etext before. The section was just plonked in ROM and sometimes due to strange alignment issues it would not get copied into RAM correctly at startup. Changed so that the section is added to ROM at a well defined address.